backend/requirements.txt
# VapeERP 백엔드 패키지

# 웹 프레임워크
fastapi==0.111.0
uvicorn[standard]==0.30.1

# DB
sqlalchemy==2.0.30
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# 인증
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# 환경변수
python-dotenv==1.0.1
pydantic-settings==2.3.4

# 데이터 처리 (seed 스크립트용)
pandas==2.2.2
openpyxl==3.1.3
